Fix logic for creating symlinks in hbb

+++ openmm/build.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 # Fix hbb issues
-ln -s /opt/rh/devtoolset-2/root/usr/lib/gcc/x86_64-CentOS-linux/ /opt/rh/devtoolset-2/root/usr/lib/gcc/x86_64-redhat-linux
-ln -s /opt/rh/devtoolset-2/root/usr/include/c++/4.8.2/x86_64-CentOS-linux/ /opt/rh/devtoolset-2/root/usr/include/c++/4.8.2/x86_64-redhat-linux
+if [ ! -e /opt/rh/devtoolset-2/root/usr/lib/gcc/x86_64-CentOS-linux/ ]; then
+    ln -s /opt/rh/devtoolset-2/root/usr/lib/gcc/x86_64-CentOS-linux/ /opt/rh/devtoolset-2/root/usr/lib/gcc/x86_64-redhat-linux
+fi
+if [ ! -e /opt/rh/devtoolset-2/root/usr/include/c++/4.8.2/x86_64-CentOS-linux/ ]; then
+    ln -s /opt/rh/devtoolset-2/root/usr/include/c++/4.8.2/x86_64-CentOS-linux/ /opt/rh/devtoolset-2/root/usr/include/c++/4.8.2/x86_64-redhat-linux
+fi
 
 # holy build box paths
 export HBB_PREFIX="/hbb_shlib"
